requirements.txt
flask==3.1.0
anthropic==0.52.0
python-dotenv==1.1.0
requests==2.32.3
beautifulsoup4==4.13.3
feedparser==6.0.11
apscheduler==3.11.0
markupsafe==3.0.2
